==> analysis/5-plot-SupplementaryFigures.Rmd <==
---
title: "5-plot SupplementaryFigures"
author: "bernard-liew"
date: "2020-06-09"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: console
---

## Load library

```{r message=FALSE, warning=FALSE}

rm (list = ls())

library(tidyverse)
library (cowplot)

```

## Plot EMG

```{r message=FALSE, warning=FALSE}
df.plot <- readRDS("output/emg_recon.RDS")

######################### Plot muscle activations ########################################
p <- emg_compare %>%
  ggplot () +
  geom_line(aes(x = cycle, y = Mean, colour = grp, linetype = type), size = 1.5) +
  scale_color_manual(values = c("black", "blue", "red")) + 
  ylab  ("% of intra-trial max EMG") + 
  xlab ("% cycle") +
  facet_wrap(muscle ~ task ,  ncol = 4, scales = "fixed") +
  theme_half_open() + 
  theme(text = element_text(size=20)) + 
  labs (colour = "Group",
        linetype = "Type") +
  guides (fill = FALSE)


p
# tiff("../../fig/suppl_fig_ms.tiff", width = 15, height = 15, units = 'in', res = 300, compression = 'none')
# p
# dev.off()


```


## Plot kinematics

```{r}

######################### Plot kinematics ########################################

kin.ord <-  c("PELVIC","RACR")

vert <- "vert"


kin.plot <- df.plot [grepl ("DISPL", names (df.plot))] %>%
  bind_rows(.id = "var") %>%
  mutate (phase = ifelse (grepl ("LIFT", var), "lift", "lower"),
          region = str_replace_all(var, c("_.*"), ""),
          region = ifelse (region == "Mid", "Mid_Shd", region),
          direction = ifelse(grepl ("AP", var), "ap", "vert")) %>%
  select (subj, grp, phase, region, direction, everything ()) %>%
  filter (region %in% kin.ord) %>%
  mutate (phase = factor (phase, levels = c ("lift", "lower")),
          region = factor (region, levels = kin.ord, labels = c("Pelvis", "Trunk") ),
          direction = factor (direction, levels = c("ap", "vert")),
          grp = factor (grp, levels = c("con", "rlbp", "clbp"), labels = c("con", "rLBP", "cLBP"))) %>%
  gather (c(V1:V101), key = cycle, value = displ) %>%
  mutate (cycle = str_remove(cycle, "V") %>% as.numeric ()) %>%
  group_by(grp, phase, region, phase, direction, cycle) %>%
  summarise(displ = mean (displ)) %>%
  ungroup()

# Vertical direction plot

p1 <- kin.plot %>%
  filter (direction == "vert") %>%
  ggplot () +
  geom_line (aes (x = cycle, y = displ, colour = grp)) +
  facet_wrap(region ~ phase, ncol = 6, scales = "free_y") +
  scale_colour_manual(values = c("black", "blue", "red")) +
  labs (x = "Cycle (0-100%)",
        y = "Displacement (m)",
        colour = "Group") +
  # Manually create values to expand the scale, by finding "pretty" 
  # values that are slightly larger than the range of y-axis values 
  # within each facet; set alpha = 0 since they aren't meant to be seen
  geom_point(data = . %>% 
               group_by(region , phase) %>% #group by facet variable
               summarise(y.min = pretty(displ)[1],
                         y.max = pretty(displ)[length(pretty(displ))]) %>%
               tidyr::gather(key, value, -region, -phase), 
             aes(x = 1, y = value),
             inherit.aes = FALSE, alpha = 0) +
  # Turn off automatical scale expansion, & manually set scale breaks
  # as an evenly spaced sequence (with the "pretty" values created above
  # providing the limits for each facet). If there are many facets to
  # show, I recommend no more than 3 labels in each facet, to keep things
  # simple.
  scale_y_continuous(breaks = function(x) seq(from = x[1], 
                                              to = x[2], 
                                              length.out = 3), 
                     expand = c(0, 0),
                     labels = scales::number_format(accuracy = 0.01)) +
  theme_half_open()

# AP direction plot

p2 <- kin.plot %>%
  filter (direction == "ap") %>%
  ggplot () +
  geom_line (aes (x = cycle, y = displ, colour = grp)) +
  facet_wrap(region ~ phase, ncol = 6, scales = "free_y") +
  scale_colour_manual(values = c("black", "blue", "red")) +
  labs (x = "Cycle (0-100%)",
        y = "Displacement (m)",
        colour = "Group") +
  # Manually create values to expand the scale, by finding "pretty" 
  # values that are slightly larger than the range of y-axis values 
  # within each facet; set alpha = 0 since they aren't meant to be seen
  geom_point(data = . %>% 
               group_by(region , phase) %>% #group by facet variable
               summarise(y.min = pretty(displ)[1],
                         y.max = pretty(displ)[length(pretty(displ))]) %>%
               tidyr::gather(key, value, -region, -phase), 
             aes(x = 1, y = value),
             inherit.aes = FALSE, alpha = 0) +
  # Turn off automatical scale expansion, & manually set scale breaks
  # as an evenly spaced sequence (with the "pretty" values created above
  # providing the limits for each facet). If there are many facets to
  # show, I recommend no more than 3 labels in each facet, to keep things
  # simple.
  scale_y_continuous(breaks = function(x) seq(from = x[1], 
                                              to = x[2], 
                                              length.out = 3), 
                     expand = c(0, 0),
                     labels = scales::number_format(accuracy = 0.01))+
  theme_half_open()


legend <- get_legend(p1+
                       guides(color = guide_legend(nrow = 1)) +
                       theme(legend.direction = "horizontal",legend.justification="center" ,legend.box.just = "bottom"))

p <- cowplot::plot_grid(p1+ theme(legend.position="none", axis.title.x = element_blank()),
                        p2+ theme(legend.position="none"), 
                        labels = c("a", "b"),
                        hjust = -1,
                        nrow = 2)

p <-  plot_grid(p, legend, ncol = 1, rel_heights = c(1, .1))

tiff("../../fig/suppl_fig_kine.tiff", width = 7, height = 5, units = 'in', res = 200, compression = 'none')
p
dev.off()


```
